moved TeamsTable into UserDashboard folder
