+++ frontend/src/components/pages/UserDashboard/UserDetails/PermissionsTable/index.js
