+++ frontend/src/components/pages/UserDashboard/UserDetails/TeamsTable/index.js
