+++ frontend/src/stories/TeamsTable.stories.js
@@ -1,11 +1,11 @@
 import React from "react";
-import TeamsTable from "../components/regions/UserDetails/TeamsTable/Presentation";
+import TeamsTable from "../components/pages/UserDashboard/UserDetails/TeamsTable/Presentation";
 import authUser from "./fixtures/authUser.json";
 import user from "./fixtures/user.json";
 import { BrowserRouter as Router } from "react-router-dom";
 
 export default {
-    title: "Tilde/UserTeamsTable",
+    title: "Tilde/pages/UserDashboard/UserTeamsTable",
     component: TeamsTable,
 }
